(WIP) split out guts of MetricBatchJson into two classes implementing two interfaces
