+++ metrics/src/main/java/com/newrelic/telemetry/json/MetricBatchJsonCommonBlockWriter.java
@@ -1,0 +1,26 @@
+package com.newrelic.telemetry.json;
+
+import com.newrelic.telemetry.Telemetry;
+import com.newrelic.telemetry.TelemetryBatch;
+
+public class MetricBatchJsonCommonBlockWriter implements JsonCommonBlockWriter {
+
+  private final AttributesJson attributesJson;
+
+  public MetricBatchJsonCommonBlockWriter(AttributesJson attributesJson) {
+    this.attributesJson = attributesJson;
+  }
+
+  @Override
+  public <T extends Telemetry> void appendCommonJson(
+      TelemetryBatch<T> batch, StringBuilder builder) {
+    if (batch.hasCommonAttributes()) {
+      builder
+          .append("\"common\":")
+          .append("{")
+          .append("\"attributes\":")
+          .append(attributesJson.toJson(batch.getCommonAttributes().asMap()))
+          .append("}");
+    }
+  }
+}
