+++ telemetry/src/test/java/com/newrelic/telemetry/TelemetryBatchJsonTest.java
@@ -1,12 +1,12 @@
-package com.newrelic.telemetry.json;
+package com.newrelic.telemetry;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import com.newrelic.telemetry.Attributes;
-import com.newrelic.telemetry.Count;
-import com.newrelic.telemetry.Metric;
-import com.newrelic.telemetry.Telemetry;
-import com.newrelic.telemetry.TelemetryBatch;
+import com.newrelic.telemetry.json.JsonCommonBlockWriter;
+import com.newrelic.telemetry.json.JsonTelemetryBlockWriter;
+import com.newrelic.telemetry.json.TelemetryBatchJson;
+import com.newrelic.telemetry.json.TypeDispatchingJsonCommonBlockWriter;
+import com.newrelic.telemetry.json.TypeDispatchingJsonTelemetryBlockWriter;
 import java.util.Collection;
 import java.util.Collections;
 import org.junit.jupiter.api.Test;
@@ -45,8 +45,9 @@ class TelemetryBatchJsonTest {
         };
     String expectedJson = "[{" + commonBit + "," + telemetryBit + "}]";
 
-    TelemetryBatchJson testClass = new TelemetryBatchJson(commonWriter, mainBodyWriter,
-        commonBlockSpanWriter, mainBodySpanWriter);
+    TelemetryBatchJson testClass = new TelemetryBatchJson(
+        new TypeDispatchingJsonCommonBlockWriter(commonWriter, null),
+        new TypeDispatchingJsonTelemetryBlockWriter(mainBodyWriter, null));
 
     String result = testClass.toJson(batch);
     assertEquals(expectedJson, result);
